move: hooks into entities
