+++ apps/web/src/components/ui/Schedule/ScheduleDisplay.tsx
@@ -1,7 +1,7 @@
 import type { Schedule } from 'fitness-struct'
 import type { ReactNode } from 'react';
 import { useGetExerciseListByScheduleId } from '@entities/exercise/api'
-import useScheduleMenu from '@hooks/useScheduleMenu'
+import { useScheduleMenu } from '@entities/schedule/hooks';
 import MenuableAccordion from '@shared/ui/MenuableAccordion';
 import SimpleFitnessList from '@ui/Fitness/SimpleFitnessList'
