+++ apps/web/src/components/ui/Schedule/ScheduleList.tsx
@@ -2,8 +2,8 @@ import { Button } from '@heroui/react';
 import { Suspense, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useGetScheduleByDate } from '@entities/schedule/api';
+import { useScheduleActions } from '@entities/schedule/hooks';
 import { ScheduleType } from '@entities/schedule/model/ScheduleType';
-import { useScheduleActions } from '@hooks/useScheduleMenu';
 import MenuableAccordion from '@shared/ui/MenuableAccordion';
 import StateRender from '@shared/ui/StateRender';
 import ScheduleDisplay from './ScheduleDisplay';
