+++ apps/web/src/entities/schedule/hooks/index.ts
@@ -1,3 +1,3 @@
 export { default as useScheduleActions } from './useScheduleActions'
 export { default as useScheduleHeaderMenu } from './useScheduleHeaderMenu'
-export { default } from './useScheduleMenu'
+export { default as useScheduleMenu } from './useScheduleMenu'
