+++ apps/web/src/entities/schedule/hooks/useScheduleMenu.ts
