+++ apps/web/src/pages/:selectDate/workout/:id.tsx
@@ -5,8 +5,8 @@ import { startTransition, Suspense, useCallback, useEffect, useMemo, useState } 
 import { useTranslation } from 'react-i18next'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useGetScheduleById, useUpdateSchedule } from '@entities/schedule/api'
+import { useScheduleHeaderMenu } from '@entities/schedule/hooks'
 import { ScheduleType } from '@entities/schedule/model/ScheduleType'
-import { useScheduleHeaderMenu } from '@hooks/useScheduleMenu'
 import usePageTracker from '@shared/hooks/usePageTracker'
 import { dayjs } from '@shared/lib/utils'
 import ExerciseDataList from '@ui/ExerciseData/ExerciseDataList'
