+++ apps/web/src/pages/Main.tsx
@@ -3,7 +3,7 @@ import { Suspense, useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate } from 'react-router-dom'
 import { useGetScheduleByDate } from '@entities/schedule/api'
-import { useScheduleActions } from '@hooks/useScheduleMenu'
+import { useScheduleActions } from '@entities/schedule/hooks'
 import usePageTracker from '@shared/hooks/usePageTracker'
 import MenuableAccordion from '@shared/ui/MenuableAccordion'
 import StateRender from '@shared/ui/StateRender'
