+++ apps/web/src/pages/preset/:id.tsx
@@ -2,8 +2,8 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useGetExercisePresetWithListById } from '@entities/exercisePreset/api'
+import { usePresetMenu } from '@entities/exercisePreset/hooks'
 import { useUpdateExerciseListByExercisePreset } from '@features/presetManagement/api'
-import usePresetMenu from '@hooks/usePresetMenu/usePresetMenu'
 import FitnessListEditor from '@ui/Fitness/FitnessListEditor'
 import { useHeaderHandler, useHeaderMenuHandler } from '@widgets/header'
